app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
gitpython
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-qdrant
langchain-ollama
qdrant-client
sentence-transformers
motor
pymongo
pyjwt